backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
transformers>=4.40.0
torch>=2.2.0
Pillow>=10.3.0
sentencepiece>=0.2.0
